backend/requirements.txt
asgiref==3.11.0
 Django>=5.2,<6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
 djangorestframework-simplejwt
 django-filter
sqlparse==0.5.5
tzdata==2025.3
Pillow
reportlab==4.2.5
